Add variable registration to ConfigVariableReader
  - Add RegisteredConfigVariable, a struct that captures a variable's
    key, default content, secrecy, and metadata for storage in
    homogeneous collections
  - Add encode closure to ConfigVariableContent for converting values
    back to ConfigContent, enabling default value encoding at
    registration time
  - Add encodeToContent on CodableValueRepresentation for converting
    encoded Data into the appropriate ConfigContent
  - Add ConfigVariableReader.register, which encodes and stores a
    variable's registration with duplicate-key detection and logging
  - Split ConfigVariableReaderTests into per-concern test files:
    scalar, array, raw representable, config expression, data
    representation, codable, and registration
  - Extract shared test mocks and metadata keys into Testing Support
