+++ Tests/DevConfigurationTests/Testing Support/MockCodableConfig.swift
@@ -1,0 +1,11 @@
+//
+//  MockCodableConfig.swift
+//  DevConfiguration
+//
+//  Created by Prachi Gauriar on 3/5/2026.
+//
+
+struct MockCodableConfig: Codable, Hashable, Sendable {
+    let variant: String
+    let count: Int
+}
