+++ Tests/DevConfigurationTests/Testing Support/MockIntEnum.swift
@@ -1,0 +1,13 @@
+//
+//  MockIntEnum.swift
+//  DevConfiguration
+//
+//  Created by Prachi Gauriar on 3/5/2026.
+//
+
+enum MockIntEnum: Int, CaseIterable, Sendable {
+    case one = 1
+    case two = 2
+    case three = 3
+    case four = 4
+}
