+++ Tests/DevConfigurationTests/Testing Support/MockStringEnum.swift
@@ -1,0 +1,13 @@
+//
+//  MockStringEnum.swift
+//  DevConfiguration
+//
+//  Created by Prachi Gauriar on 3/5/2026.
+//
+
+enum MockStringEnum: String, CaseIterable, Sendable {
+    case alpha
+    case bravo
+    case charlie
+    case delta
+}
