+++ Tests/DevConfigurationTests/Testing Support/UnencodableValue.swift
@@ -1,0 +1,8 @@
+//
+//  UnencodableValue.swift
+//  DevConfiguration
+//
+//  Created by Prachi Gauriar on 3/5/2026.
+//
+
+struct UnencodableValue: Sendable {}
