+++ Tests/DevConfigurationTests/Unit Tests/Core/ConfigVariableMetadataTests.swift
@@ -114,41 +114,3 @@ struct ConfigVariableMetadataTests: RandomValueGenerating {
         #expect(OptionalEnumMetadataKey.displayText(for: nil) == nil)
     }
 }
-
-
-// MARK: - Test Metadata Keys
-
-private enum MetadataEnum: String, CaseIterable, Sendable {
-    case valueA
-    case valueB
-}
-
-
-private struct EnumMetadataKey: ConfigVariableMetadataKey {
-    static let defaultValue = MetadataEnum.valueA
-    static let keyDisplayText = "EnumKey"
-}
-
-
-private struct IntMetadataKey: ConfigVariableMetadataKey {
-    static let defaultValue = 0
-    static let keyDisplayText = "IntKey"
-}
-
-
-private struct OptionalEnumMetadataKey: ConfigVariableMetadataKey {
-    static let defaultValue: MetadataEnum? = nil
-    static let keyDisplayText = "OptionalEnumKey"
-}
-
-
-private struct OptionalIntMetadataKey: ConfigVariableMetadataKey {
-    static let defaultValue: Int? = nil
-    static let keyDisplayText = "OptionalIntKey"
-}
-
-
-private struct StringMetadataKey: ConfigVariableMetadataKey {
-    static let defaultValue: String? = nil
-    static let keyDisplayText = "StringKey"
-}
