+++ Tests/DevConfigurationTests/Unit Tests/Core/RegisteredConfigVariableTests.swift
@@ -1,0 +1,50 @@
+//
+//  RegisteredConfigVariableTests.swift
+//  DevConfiguration
+//
+//  Created by Prachi Gauriar on 3/5/2026.
+//
+
+import Configuration
+import DevTesting
+import Testing
+
+@testable import DevConfiguration
+
+struct RegisteredConfigVariableTests: RandomValueGenerating {
+    var randomNumberGenerator = makeRandomNumberGenerator()
+
+
+    @Test
+    mutating func dynamicMemberLookupReturnsMetadataValue() {
+        // set up
+        var metadata = ConfigVariableMetadata()
+        let project = randomAlphanumericString()
+        metadata[TestProjectMetadataKey.self] = project
+
+        let variable = RegisteredConfigVariable(
+            key: randomConfigKey(),
+            defaultContent: randomConfigContent(),
+            secrecy: randomConfigVariableSecrecy(),
+            metadata: metadata
+        )
+
+        // expect
+        #expect(variable.testProject == project)
+    }
+
+
+    @Test
+    mutating func dynamicMemberLookupReturnsDefaultWhenNotSet() {
+        // set up
+        let variable = RegisteredConfigVariable(
+            key: randomConfigKey(),
+            defaultContent: randomConfigContent(),
+            secrecy: randomConfigVariableSecrecy(),
+            metadata: ConfigVariableMetadata()
+        )
+
+        // expect
+        #expect(variable.testProject == nil)
+    }
+}
